backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
requests>=2.32
python-dotenv>=1.1
